server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.1